refactor(network): move evpn cmd files in network dir
